add database for calories that appears alongside steps in our Stats Activity

+++ app/src/main/java/com/mikeschen/www/fitnessapp/Calories.java
@@ -5,24 +5,24 @@ package com.mikeschen.www.fitnessapp;
  */
 public class Calories {
 
-    int id;
+    long id;
     int caloriesBurned;
     int date;
 
     public Calories() {
     }
 
-    public Calories(int id, int caloriesBurned, int date) {
+    public Calories(long id, int caloriesBurned, int date) {
         this.id = id;
         this.caloriesBurned = caloriesBurned;
         this.date = date;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
